requirements.txt
Flask==2.1.1
Werkzeug==2.0.2
Flask-PyMongo==2.3.0
python-dotenv==0.19.1